\name{midi.scale}
\alias{midi.scale}
\alias{midi.scales.adonai_malakh}
\alias{midi.scales.aeolian_flat_0}
\alias{midi.scales.aeolian_natural_minor_asavari_a_to_g}
\alias{midi.scales.algerian}
\alias{midi.scales.altered_pentatonic}
\alias{midi.scales.alternating_tetramirror}
\alias{midi.scales.anhemitonic_hexatonic}
\alias{midi.scales.augmented_chord}
\alias{midi.scales.avaha}
\alias{midi.scales.balinese}
\alias{midi.scales.balinese_pentachord}
\alias{midi.scales.bartok}
\alias{midi.scales.bhairavi_e_to_d_asc_nats}
\alias{midi.scales.bilaval_c_to_b_asc_nats}
\alias{midi.scales.blues}
\alias{midi.scales.blues_dorian_hexatonic}
\alias{midi.scales.blues_enneatonic}
\alias{midi.scales.blues_heptatonic}
\alias{midi.scales.blues_minor}
\alias{midi.scales.blues_minor_maj6}
\alias{midi.scales.blues_modified}
\alias{midi.scales.blues_octatonic}
\alias{midi.scales.blues_pentacluster}
\alias{midi.scales.blues_pentatonic_minor}
\alias{midi.scales.blues_phrygian}
\alias{midi.scales.blues_sharp_v}
\alias{midi.scales.blues_with_leading_tone}
\alias{midi.scales.center_cluster_pentamirror}
\alias{midi.scales.chaio}
\alias{midi.scales.chinese}
\alias{midi.scales.chromatic_bebop}
\alias{midi.scales.chromatic_decamirror}
\alias{midi.scales.chromatic_diatonic_dorian}
\alias{midi.scales.chromatic_dorian}
\alias{midi.scales.chromatic_dorian_inverse}
\alias{midi.scales.chromatic_heptamirror}
\alias{midi.scales.chromatic_hexamirror}
\alias{midi.scales.chromatic_hypodorian}
\alias{midi.scales.chromatic_hypodorian_inverse}
\alias{midi.scales.chromatic_hypolydian}
\alias{midi.scales.chromatic_hypolydian_inverse}
\alias{midi.scales.chromatic_hypophrygian_inverse}
\alias{midi.scales.chromatic_lydian}
\alias{midi.scales.chromatic_lydian_inverse}
\alias{midi.scales.chromatic_mixolydian}
\alias{midi.scales.chromatic_mixolydian_inverse}
\alias{midi.scales.chromatic_nonamirror}
\alias{midi.scales.chromatic_octamirror}
\alias{midi.scales.chromatic_pentamirror}
\alias{midi.scales.chromatic_permuted_diatonic_dorian}
\alias{midi.scales.chromatic_phrygian}
\alias{midi.scales.chromatic_phrygian_inverse}
\alias{midi.scales.chromatic_tetramirror}
\alias{midi.scales.chromatic_trimirror}
\alias{midi.scales.chromatic_undecamirror_0}
\alias{midi.scales.damian_emmanuel}
\alias{midi.scales.diminished}
\alias{midi.scales.diminished_6th_chord}
\alias{midi.scales.diminished_chord}
\alias{midi.scales.diminished_whole_tone}
\alias{midi.scales.dominant_bebop}
\alias{midi.scales.dominant_pentatonic}
\alias{midi.scales.dorian}
\alias{midi.scales.dorian_aeolian}
\alias{midi.scales.dorian_b4}
\alias{midi.scales.dorian_bii}
\alias{midi.scales.dorian_pentatonic}
\alias{midi.scales.dorian_tetrachord}
\alias{midi.scales.dorico_flamenco}
\alias{midi.scales.double_phrygian_hexatonic}
\alias{midi.scales.egyptian}
\alias{midi.scales.enigmatic_major}
\alias{midi.scales.enigmatic_minor}
\alias{midi.scales.eskimo_hexatonic_1}
\alias{midi.scales.eskimo_tetratonic}
\alias{midi.scales.ethiopian}
\alias{midi.scales.example_of_pelog}
\alias{midi.scales.example_of_slendro}
\alias{midi.scales.flamenco}
\alias{midi.scales.four_semitones}
\alias{midi.scales.full_minor}
\alias{midi.scales.genus_chromaticum}
\alias{midi.scales.genus_secundum}
\alias{midi.scales.gipsy_hexatonic}
\alias{midi.scales.gnossiennes}
\alias{midi.scales.half_diminished}
\alias{midi.scales.half_diminished_bebop}
\alias{midi.scales.han_kumo}
\alias{midi.scales.hard_japan}
\alias{midi.scales.harmonic_major_0}
\alias{midi.scales.harmonic_major_1}
\alias{midi.scales.harmonic_minor}
\alias{midi.scales.harmonic_minor_inverse_or_bhairav}
\alias{midi.scales.harmonic_minor_tetrachord}
\alias{midi.scales.harmonic_neapolitan_minor}
\alias{midi.scales.hawaiian}
\alias{midi.scales.hindi}
\alias{midi.scales.hindi_2_flats_and_bv}
\alias{midi.scales.hindi_4_flats_biv}
\alias{midi.scales.hindi_bii_biii_bvii}
\alias{midi.scales.hindi_sharp_iv_and_sharp_v}
\alias{midi.scales.hindolita}
\alias{midi.scales.hindu_bvi_and_bvii}
\alias{midi.scales.hirajoshi_japan}
\alias{midi.scales.honchoshi_plagal_form}
\alias{midi.scales.houseini}
\alias{midi.scales.houzam}
\alias{midi.scales.hungarian_folk_or_byzantine}
\alias{midi.scales.hungarian_gypsy}
\alias{midi.scales.hungarian_major}
\alias{midi.scales.hungarian_minor}
\alias{midi.scales.hungarian_minor_b1}
\alias{midi.scales.indian}
\alias{midi.scales.ionian}
\alias{midi.scales.ionian_pentatonic}
\alias{midi.scales.ionian_sharp_4}
\alias{midi.scales.iwato_japan}
\alias{midi.scales.japanese}
\alias{midi.scales.japanese_pentachord}
\alias{midi.scales.javanese}
\alias{midi.scales.jazz_minor_inverse}
\alias{midi.scales.jg_octatonic}
\alias{midi.scales.kafi_d_to_c_asc_naturals}
\alias{midi.scales.kalyan_f_to_e_asc_naturals}
\alias{midi.scales.khamaj_g_to_f_asc_nats}
\alias{midi.scales.kiourdi}
\alias{midi.scales.kokin_joshi}
\alias{midi.scales.kumoi_japan}
\alias{midi.scales.kung}
\alias{midi.scales.leading_whole_tone}
\alias{midi.scales.locrian_b_to_a_asc_naturals}
\alias{midi.scales.locrian_bb6}
\alias{midi.scales.locrian_natural_maj_5}
\alias{midi.scales.locrian_pentamirror}
\alias{midi.scales.lydian}
\alias{midi.scales.lydian_augmented}
\alias{midi.scales.lydian_diminished}
\alias{midi.scales.lydian_dominant}
\alias{midi.scales.lydian_hexatonic}
\alias{midi.scales.lydian_minor}
\alias{midi.scales.lydian_mixolydian}
\alias{midi.scales.lydian_pentachord}
\alias{midi.scales.lydian_sharp_1}
\alias{midi.scales.lydian_sharp_1_hexatonic}
\alias{midi.scales.magen_abot}
\alias{midi.scales.major}
\alias{midi.scales.major_6_b4_chord}
\alias{midi.scales.major_6_chord}
\alias{midi.scales.major_and_minor_mixed}
\alias{midi.scales.major_b6_chord}
\alias{midi.scales.major_b6_suspend_3th_chord}
\alias{midi.scales.major_bebop}
\alias{midi.scales.major_bebop_heptatonic}
\alias{midi.scales.major_blues}
\alias{midi.scales.major_chord}
\alias{midi.scales.major_locrian}
\alias{midi.scales.major_lydian}
\alias{midi.scales.major_minor}
\alias{midi.scales.major_pentachord}
\alias{midi.scales.major_pentatonic}
\alias{midi.scales.major_suspended_3th_chord}
\alias{midi.scales.major_tetrachord}
\alias{midi.scales.maqam_hijaz}
\alias{midi.scales.maqam_shadd_araban}
\alias{midi.scales.marva_or_marvi}
\alias{midi.scales.messiaen_mode_2}
\alias{midi.scales.messiaen_mode_2_inverse}
\alias{midi.scales.messiaen_mode_3}
\alias{midi.scales.messiaen_mode_3_inverse}
\alias{midi.scales.messiaen_mode_4}
\alias{midi.scales.messiaen_mode_4_inverse}
\alias{midi.scales.messiaen_mode_5}
\alias{midi.scales.messiaen_mode_5_inverse}
\alias{midi.scales.messiaen_mode_6}
\alias{midi.scales.messiaen_mode_6_inverse}
\alias{midi.scales.messiaen_truncated_mode}
\alias{midi.scales.messiaen_truncated_mode_1}
\alias{midi.scales.messiaen_truncated_mode_2_augumented}
\alias{midi.scales.messiaen_truncated_mode_2_half_augumented}
\alias{midi.scales.messiaen_truncated_mode_4}
\alias{midi.scales.messiaen_truncated_mode_4_inverse}
\alias{midi.scales.messiaen_truncated_mode_5}
\alias{midi.scales.minor_5th_added}
\alias{midi.scales.minor_b6_chord}
\alias{midi.scales.minor_bebop}
\alias{midi.scales.minor_bebop_heptatonic}
\alias{midi.scales.minor_locrian}
\alias{midi.scales.minor_melodic}
\alias{midi.scales.minor_pentachord}
\alias{midi.scales.minor_pentatonic_with_leading_tones}
\alias{midi.scales.mixolydian}
\alias{midi.scales.mixolydian_augmented}
\alias{midi.scales.mixolydian_b4}
\alias{midi.scales.mixolydian_hexatonic}
\alias{midi.scales.mixolydian_pentatonic}
\alias{midi.scales.moorish_phrygian}
\alias{midi.scales.nat_bhairava}
\alias{midi.scales.neapolitan_major}
\alias{midi.scales.neapolitan_minor}
\alias{midi.scales.neapolitan_minor_mode}
\alias{midi.scales.neveseri}
\alias{midi.scales.nohkan}
\alias{midi.scales.one_and_three_tones}
\alias{midi.scales.oriental}
\alias{midi.scales.oriental_0}
\alias{midi.scales.oriental_1}
\alias{midi.scales.oriental_pentacluster}
\alias{midi.scales.overtone}
\alias{midi.scales.pelog_0}
\alias{midi.scales.pelog_1}
\alias{midi.scales.persian}
\alias{midi.scales.phrygian}
\alias{midi.scales.phrygian_aeolian}
\alias{midi.scales.phrygian_dim_3th}
\alias{midi.scales.phrygian_dominant}
\alias{midi.scales.phrygian_hexatonic}
\alias{midi.scales.phrygian_locrian}
\alias{midi.scales.phrygian_major}
\alias{midi.scales.phrygian_tetrachord}
\alias{midi.scales.prokofiev}
\alias{midi.scales.prometheus}
\alias{midi.scales.prometheus_neapolitan}
\alias{midi.scales.pseudo_turkish}
\alias{midi.scales.puravi_bvi}
\alias{midi.scales.purvi_bvii}
\alias{midi.scales.pyramid_hexatonic}
\alias{midi.scales.raga_abhogi}
\alias{midi.scales.raga_bagesri}
\alias{midi.scales.raga_chandrakauns_kafi}
\alias{midi.scales.raga_desh}
\alias{midi.scales.raga_gandharavam}
\alias{midi.scales.raga_harikauns}
\alias{midi.scales.raga_jayakauns}
\alias{midi.scales.raga_jyoti}
\alias{midi.scales.raga_kalakanti}
\alias{midi.scales.raga_khamaji_durga}
\alias{midi.scales.raga_latika}
\alias{midi.scales.raga_madhuri}
\alias{midi.scales.raga_malkauns}
\alias{midi.scales.raga_megh}
\alias{midi.scales.raga_mian_ki_malhar}
\alias{midi.scales.raga_nata}
\alias{midi.scales.raga_paraju}
\alias{midi.scales.raga_rukmangi}
\alias{midi.scales.raga_shri_kalyan}
\alias{midi.scales.raga_tilang}
\alias{midi.scales.raga_trimurti}
\alias{midi.scales.raga_vaijayanti}
\alias{midi.scales.raga_valaji}
\alias{midi.scales.raga_vasanta}
\alias{midi.scales.raga_vasantabhairavi}
\alias{midi.scales.raga_vijayanagari}
\alias{midi.scales.raga_vijayasri}
\alias{midi.scales.raga_vijayavasanta}
\alias{midi.scales.raga_viyogavarali}
\alias{midi.scales.raga_vutari}
\alias{midi.scales.raga_yamuna_kalyani}
\alias{midi.scales.raga_zilaf}
\alias{midi.scales.ritsu}
\alias{midi.scales.ritusen_japan}
\alias{midi.scales.rock_n_roll}
\alias{midi.scales.romanian}
\alias{midi.scales.romanian_bacovia}
\alias{midi.scales.romanian_major}
\alias{midi.scales.sabach}
\alias{midi.scales.sakura_pentatonic}
\alias{midi.scales.scottish_hexatonic}
\alias{midi.scales.scottish_pentatonic}
\alias{midi.scales.scriabin}
\alias{midi.scales.shostakovich}
\alias{midi.scales.soft_asc_japan}
\alias{midi.scales.soft_desc_japan}
\alias{midi.scales.spanish_7_tones}
\alias{midi.scales.spanish_heptatonic}
\alias{midi.scales.spanish_pentacluster}
\alias{midi.scales.spanish_phrygian}
\alias{midi.scales.stravinski}
\alias{midi.scales.superlocrian}
\alias{midi.scales.suspended_pentatonic}
\alias{midi.scales.symmetric}
\alias{midi.scales.symmetrical_nonatonic}
\alias{midi.scales.takemitsu_tree_line_mode_0}
\alias{midi.scales.takemitsu_tree_line_mode_1}
\alias{midi.scales.three_semitones}
\alias{midi.scales.three_tones_and_one}
\alias{midi.scales.todi_bvi}
\alias{midi.scales.todi_bvii}
\alias{midi.scales.ultra_locrian}
\alias{midi.scales.untitled_nonatonic_0}
\alias{midi.scales.untitled_nonatonic_1}
\alias{midi.scales.utility_minor}
\alias{midi.scales.verdi_enigmatic}
\alias{midi.scales.verdi_enigmatic_asc}
\alias{midi.scales.verdi_enigmatic_desc}
\alias{midi.scales.warao_tetratonic}
\alias{midi.scales.whole_tone}
\alias{midi.scales.whole_tone_alternate}
\alias{midi.scales.whole_tone_tetramirror}
\alias{midi.scales.youlan}
\alias{midi.scales.zirafkend}
\alias{midi.scales.chromatic}
\alias{midi.notes.cb}
\alias{midi.notes.c}
\alias{midi.notes.cs}
\alias{midi.notes.db}
\alias{midi.notes.d}
\alias{midi.notes.ds}
\alias{midi.notes.eb}
\alias{midi.notes.e}
\alias{midi.notes.es}
\alias{midi.notes.fb}
\alias{midi.notes.f}
\alias{midi.notes.fs}
\alias{midi.notes.gb}
\alias{midi.notes.g}
\alias{midi.notes.gs}
\alias{midi.notes.ab}
\alias{midi.notes.a}
\alias{midi.notes.as}
\alias{midi.notes.bb}
\alias{midi.notes.b}
\alias{midi.notes.bs}

\title{Apply a Scale to a Pitch Vector}
\description{
	Impose a scale on a vector of pitches by rounding to the closest pitch
	belonging to the scale.
}
\usage{
midi.scale(pitch, scale = midi.scales.chromatic)
}
\arguments{
	\item{pitch}{a vector of pitches}
	\item{scale}{a vector of integers, 0 <= scale <= 11}
}
\details{
	midi.scale takes a vector of arbitrary range and length and rounds its
	element to the closest element in the scale.  The scale is defined
	within one octave, but is then modified internally to contain the full
	range of valid midi pitches (0--127).

	This means that any pitch that is higher than the highest pitch in the
	scale, in the highest octave, will be rounded down to that note.  The
	same applies for notes lower than the lowest note in the scale, in the
	lowest octave.

	There is currently no way to define a scale that varies over multiple
	octaves.

	Bundled with rmidi is a comprehensive set of scales, taken from
	\url{http://www.lucytune.com/250ScaleCoding.pdf}.
	These are all prefixed with \code{midi.scales.} and are best explored
	by hitting the TAB key.  All predefined scales start from C,
	e.g. midi.scales.major implies C major.

	rmidi also provides a set of helper variables for the note names,
	prefixed with \code{midi.notes.}.  Sharps are appended with \code{s}
	and flats	are appended with \code{f}.
}
\value{
	Returns a vector with a scale applied.
}
\references{
	Lucy Tune --- 250 + Musical Scales and Scalecoding
	\url{http://www.lucytune.com/250ScaleCoding.pdf}
}
\seealso{
	\code{\link{midi.note}}
}
\examples{
## create a custom (somewhat crazy) scale
my.scale <- c(midi.notes.c, midi.notes.e, midi.notes.f, midi.notes.ab,
              midi.notes.as, midi.notes.b)
v1 <- 0:127
scaled.v1 <- midi.scale(v1, my.scale)

## use a predefined scale
v2 <- 127:0
scaled.v2 <- midi.scale(v2, midi.scales.dorico_flamenco)
}